+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.0.8 - 2021-02-28
+~~~~~~~~~~~~~~~~~~~
+
+* Publish large messages in `frame_max`-sized chunks
+
 v0.0.7 - 2021-02-26
 ~~~~~~~~~~~~~~~~~~~